Refactor test utils definition
