+++ apps/console/test-configs/__mocks__/redux-store-state-mock.json
