+++ apps/myaccount/test-configs/utils.tsx
@@ -19,8 +19,8 @@
 import { RenderResult, render as rtlRender } from "@testing-library/react";
 import React, { ComponentType, PropsWithChildren, ReactElement } from "react";
 import { Provider } from "react-redux";
+import ReduxStoreStateMock from "./__mocks__/redux-store-state-mock.json";
 import { mockStore } from "./redux-store-mock";
-import ReduxStoreStateMock from "./redux-store-state-mock.json";
 
 /**
  * Custom render method to includes things like global context providers, data stores, etc.
