+++ configs/common.tsconfig.json
@@ -18,7 +18,6 @@
     "exclude": [
         "node_modules", 
         "test-configs",
-        "test-utils",
         "**/tests/*",
         "**/__tests__/*",
         "**/__mocks__/*",
